+++ pyproject.toml
@@ -23,10 +23,8 @@ paraffin = 'paraffin.cli:app'
 ruff = "^0.5.2"
 zntrack = "^0.8.0a5"
 pre-commit = "^3.7.1"
-dash = "^2.17.1"
 pytest = "^8.3.3"
 coverage = "^7.6.1"
-pandas = "^2.2.3"
 
 [build-system]
 requires = ["poetry-core"]
